Rob Papen Unifiers: complete rebuild
This now works by instantiating the RP plug-in and interpreting a (drag/dropped) __CatIndex.ind file to pull out preset names in a selected bank, sending MIDI program change messages to the plug-in to change the preset, then using GetStateInformation to save the resulting state.

This works, but has many limitations. Much more refinement remains to be done.