PatchReformer: add new "preserve INST1 INST plug-in" checkbox
The old version assumed that the main instrument plug-in used in the input patches is identical to that in the reference patch. This change allows using a reference patch built with a different instrument plug-in (including just a difference of format e.g. VST/VST3, or a completely different instrument).